Hack out most of calculate_allocations()
Pause here because the test cases have made me realize my designed
config file format doesn't support situations where multiple things have
the same weight.  That is, if we have a 50/50 portfolio, 50 is the dict
key for both funds, which looks nice but doesn't work because of the
hash collision.  So I need to rethink what I want the config UX to look
like.