Fix 50/50 problem
Using percentages as keys into the dictionaries looked nice in the
config, but is a problem programmatically because we can have duplicate
keys (for instance, if a portfolio has a 50/50 split).  I thought for a
while on different ways to address this, but decided the best method for
now that keeps the config almost as simple and readable is to make each
item a list of N 1-element dicts instead of an N-element dict.