Change license specification in pyproject.toml
The previous format is what's specified in the documentation.  However,
it appears to be a new thing and not actually working, despite me having
the most recent versions of the appropriate libraries - so we'll use
this whatever funky syntax instead.

I didn't notice this before because I wasn't running pip-compile.