Pick a Schwab API client and do initial auth work
There are basically three clients to choose from:

1. https://github.com/tylerebowers/Schwabdev is the most popular, by a
   bit.
2. https://github.com/alexgolec/schwab-py has better documentation, and
   what I chose.
3. https://github.com/itsjafer/schwab-api is a junior engineer's little
   project, but notably has code to run a headless browser and
   regenerate the token without user input.  That's important because
   Schwab insists the refresh token (the long one) can only live for a
   week.

I have a feeling I'll end up needing to pull some code from option three
at some point, but I'm going to build with this one for now and get
there when I need it.  If it does happen, it'll just change how we get
the initial token we pass to this client, so it shouldn't require
ripping out the whole thing or anything like that.