Start figuring out rebalancing
In my mind I'll have a number of different rebalancing methods, and we
switch between them via config and the strategy pattern.  I'll want at
least two because for taxable accounts I want to only rebalance via
contributions, so I did some of the class setup for that to happen.

It made sense to pull the client into the Account, because it's really a
Schwab-specific thing and this may in the future support multiple
brokers.  But I have a feeling as it grows I'll need to do some larger
refactoring at some point.

I know there's a bug in here because when I switch the buy sort order it
vastly changes the results.  I'm going to need to go write a bunch of
tests, and maybe pull things apart so they can be tested, to figure out
what's wrong.